+++ src/main/java/com/maxmind/minfraud/response/Device.java
@@ -25,17 +25,8 @@ public final class Device extends AbstractModel {
      * @param confidence The device confidence.
      * @param id         The device ID.
      * @param lastSeen   When the device was last seen.
-     * @deprecated This constructor only exists for backward compatibility
-     * and will be removed in the next major release.
+     * @param localTime  The local time of the device.
      */
-    public Device(
-            Double confidence,
-            UUID id,
-            String lastSeen
-    ) {
-        this(confidence, id, lastSeen, null);
-    }
-
     public Device(
             @JsonProperty("confidence") Double confidence,
             @JsonProperty("id") UUID id,