+++ src/main/java/com/maxmind/minfraud/response/Email.java
@@ -16,6 +16,13 @@ public final class Email extends AbstractModel {
     private final String firstSeen;
     private final EmailDomain domain;
 
+    /**
+     * @param domain       The {@code EmailDomain} model object.
+     * @param isDisposable Whether it is a disposable email.
+     * @param isFree       Whether it is a free email.
+     * @param isHighRisk   Whether it is a high risk email.
+     * @param firstSeen    When the email was first seen.
+     */
     public Email(
             @JsonProperty("domain") EmailDomain domain,
             @JsonProperty("is_disposable") Boolean isDisposable,
@@ -28,51 +35,6 @@ public final class Email extends AbstractModel {
         this.isFree = isFree;
         this.isHighRisk = isHighRisk;
         this.firstSeen = firstSeen;
-    }
-
-    /**
-     * @param isDisposable Whether it is a disposable email.
-     * @param isFree       Whether it is a free email.
-     * @param isHighRisk   Whether it is a high risk email.
-     * @param firstSeen    When the email was first seen.
-     * @deprecated This constructor only exists for backward compatibility
-     * and will be removed in the next major release.
-     */
-    public Email(
-            Boolean isDisposable,
-            Boolean isFree,
-            Boolean isHighRisk,
-            String firstSeen
-    ) {
-        this(null, isDisposable, isFree, isHighRisk, firstSeen);
-    }
-
-    /**
-     * @param isFree     Whether it is a free email.
-     * @param isHighRisk Whether it is a high risk email.
-     * @param firstSeen  When the email was first seen.
-     * @deprecated This constructor only exists for backward compatibility
-     * and will be removed in the next major release.
-     */
-    public Email(
-            Boolean isFree,
-            Boolean isHighRisk,
-            String firstSeen
-    ) {
-        this(null, isFree, isHighRisk, firstSeen);
-    }
-
-    /**
-     * @param isFree     Whether it is a free email.
-     * @param isHighRisk Whether it is a high risk email.
-     * @deprecated This constructor only exists for backward compatibility
-     * and will be removed in the next major release.
-     */
-    public Email(
-            Boolean isFree,
-            Boolean isHighRisk
-    ) {
-        this(null, isFree, isHighRisk, null);
     }
 
     public Email() {