+++ src/main/java/com/maxmind/minfraud/response/IpAddress.java
@@ -12,31 +12,9 @@ import java.util.List;
  * This class contains minFraud response data related to the IP location
  */
 public final class IpAddress extends InsightsResponse implements IpAddressInterface {
-    private final GeoIp2Country country;
     private final GeoIp2Location location;
     private final Double risk;
     private final List<IpRiskReason> riskReasons;
-
-    public IpAddress(
-            @JsonProperty("city") City city,
-            @JsonProperty("continent") Continent continent,
-            @JsonProperty("country") GeoIp2Country country,
-            @JsonProperty("location") GeoIp2Location location,
-            @JsonProperty("maxmind") MaxMind maxmind,
-            @JsonProperty("postal") Postal postal,
-            @JsonProperty("registered_country") Country registeredCountry,
-            @JsonProperty("represented_country") RepresentedCountry representedCountry,
-            @JsonProperty("risk") Double risk,
-            @JsonProperty("risk_reasons") List<IpRiskReason> riskReasons,
-            @JsonProperty("subdivisions") List<Subdivision> subdivisions,
-            @JsonProperty("traits") Traits traits
-    ) {
-        super(city, continent, country, location, maxmind, postal, registeredCountry, representedCountry, subdivisions, traits);
-        this.country = country == null ? new GeoIp2Country() : country;
-        this.location = location == null ? new GeoIp2Location() : location;
-        this.risk = risk;
-        this.riskReasons = Collections.unmodifiableList(riskReasons == null ? new ArrayList<>() : riskReasons);
-    }
 
     /**
      * @param city               The city information.
@@ -48,40 +26,32 @@ public final class IpAddress extends InsightsResponse implements IpAddressInterf
      * @param registeredCountry  The information about the country where the IP was registered.
      * @param representedCountry The represented country, e.g., for military bases in other countries.
      * @param risk               The IP risk.
+     * @param riskReasons        The reasons for the IP risk.
      * @param subdivisions       The list of subdivisions.
      * @param traits             Information about various other IP traits.
-     * @deprecated This constructor only exists for backward compatibility
-     * and will be removed in the next major release.
      */
     public IpAddress(
-            City city,
-            Continent continent,
-            GeoIp2Country country,
-            GeoIp2Location location,
-            MaxMind maxmind,
-            Postal postal,
-            Country registeredCountry,
-            RepresentedCountry representedCountry,
-            Double risk,
-            List<Subdivision> subdivisions,
-            Traits traits
+            @JsonProperty("city") City city,
+            @JsonProperty("continent") Continent continent,
+            @JsonProperty("country") Country country,
+            @JsonProperty("location") GeoIp2Location location,
+            @JsonProperty("maxmind") MaxMind maxmind,
+            @JsonProperty("postal") Postal postal,
+            @JsonProperty("registered_country") Country registeredCountry,
+            @JsonProperty("represented_country") RepresentedCountry representedCountry,
+            @JsonProperty("risk") Double risk,
+            @JsonProperty("risk_reasons") List<IpRiskReason> riskReasons,
+            @JsonProperty("subdivisions") List<Subdivision> subdivisions,
+            @JsonProperty("traits") Traits traits
     ) {
-        this(city, continent, country, location, maxmind, postal, registeredCountry, representedCountry,
-                risk, null, subdivisions, traits);
+        super(city, continent, country, location, maxmind, postal, registeredCountry, representedCountry, subdivisions, traits);
+        this.location = location == null ? new GeoIp2Location() : location;
+        this.risk = risk;
+        this.riskReasons = Collections.unmodifiableList(riskReasons == null ? new ArrayList<>() : riskReasons);
     }
 
     public IpAddress() {
-        this(null, null, null, null, null, null, null, null, null, null, null);
-    }
-
-    /**
-     * @return Country record for the requested IP address. This object
-     * represents the country where MaxMind believes the end user is
-     * located.
-     */
-    @Override
-    public GeoIp2Country getCountry() {
-        return country;
+        this(null, null, null, null, null, null, null, null, null, null, null, null);
     }
 
     /**