+++ src/test/java/com/maxmind/minfraud/response/IpAddressTest.java
@@ -47,7 +47,6 @@ public class IpAddressTest extends AbstractOutputTest {
         assertEquals("IP risk", Double.valueOf(99), address.getRisk());
         assertEquals("correct local time", time, address.getLocation().getLocalTime());
         assertEquals("1.2.0.0/16", address.getTraits().getNetwork().toString());
-        assertTrue("isHighRisk", address.getCountry().isHighRisk());
         assertTrue("isAnonymous", address.getTraits().isAnonymous());
         assertTrue("isAnonymousVpn", address.getTraits().isAnonymousVpn());
         assertTrue("isHostingProvider", address.getTraits().isHostingProvider());